chore: move files in log-viewer into folders
