+++ log-viewer/modules/components/LogLevels.ts
@@ -4,9 +4,9 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
 
-import { LogSetting } from '../../parsers/TreeParser.js';
-import { globalStyles } from '../../styles/global.styles.js';
-import { skeletonStyles } from '../skeleton/skeleton.styles.js';
+import { LogSetting } from '../parsers/TreeParser.js';
+import { globalStyles } from '../styles/global.styles.js';
+import { skeletonStyles } from './skeleton/skeleton.styles.js';
 
 @customElement('log-levels')
 export class LogLevels extends LitElement {
